Refactored upload and replace to use deploy
