+++ cmd/deploy_test.go
@@ -12,6 +12,44 @@ import (
 	"github.com/Shopify/themekit/src/shopify"
 )
 
+func TestUpload(t *testing.T) {
+	ctx, client, _, _, _ := createTestCtx()
+	ctx.Args = []string{"templates/layout.liquid"}
+	ctx.Flags.NoDelete = true
+	client.On("UpdateAsset", shopify.Asset{Key: "templates/layout.liquid"}).Return(nil)
+	err := deploy(ctx)
+	assert.NotNil(t, err)
+
+	ctx, client, _, _, _ = createTestCtx()
+	ctx.Args = []string{"templates/layout.liquid"}
+	ctx.Flags.NoDelete = true
+	ctx.Env.ReadOnly = true
+	client.On("UpdateAsset", shopify.Asset{Key: "templates/layout.liquid"}).Return(nil)
+	err = deploy(ctx)
+	if assert.NotNil(t, err) {
+		assert.Contains(t, err.Error(), "environment is readonly")
+	}
+
+	ctx, client, _, stdOut, _ := createTestCtx()
+	ctx.Args = []string{"assets/app.js"}
+	ctx.Flags.NoDelete = true
+	ctx.Env.Directory = "_testdata/projectdir"
+	ctx.Flags.Verbose = true
+	client.On("UpdateAsset", shopify.Asset{Key: "assets/app.js"}).Return(nil)
+	err = deploy(ctx)
+	assert.Nil(t, err)
+	assert.Contains(t, stdOut.String(), "Updated assets/app.js")
+
+	ctx, client, _, stdOut, _ = createTestCtx()
+	ctx.Env.Directory = "_testdata/projectdir"
+	ctx.Flags.Verbose = true
+	ctx.Flags.NoDelete = true
+	client.On("UpdateAsset", mock.MatchedBy(func(a shopify.Asset) bool { return true })).Return(nil)
+	err = deploy(ctx)
+	assert.Nil(t, err)
+	assert.Contains(t, stdOut.String(), "Updated config/settings_data.json")
+}
+
 func TestReplace(t *testing.T) {
 	ctx, client, _, stdOut, _ := createTestCtx()
 	ctx.Flags.Verbose = true
@@ -19,14 +57,14 @@ func TestReplace(t *testing.T) {
 	client.On("GetAllAssets").Return([]string{"assets/logo.png"}, nil)
 	client.On("UpdateAsset", mock.MatchedBy(func(shopify.Asset) bool { return true })).Return(nil).Times(2)
 	client.On("DeleteAsset", mock.MatchedBy(func(shopify.Asset) bool { return true })).Return(nil).Once()
-	err := replace(ctx)
+	err := deploy(ctx)
 	assert.Nil(t, err)
 	assert.Contains(t, stdOut.String(), "Updated config/settings_data.json")
 	client.AssertExpectations(t)
 
 	ctx, client, _, _, _ = createTestCtx()
 	client.On("GetAllAssets").Return([]string{}, fmt.Errorf("server error"))
-	err = replace(ctx)
+	err = deploy(ctx)
 	if assert.NotNil(t, err) {
 		assert.Contains(t, err.Error(), "server error")
 	}
